Split resource/asset.go into resource/asset and resource/archive
This PR is motivated by https://github.com/pulumi/pulumi/pull/15145.

`resource.*` should be built on top of `property.Value`, which means that `resource`
needs to be able to import `property.Value`, and so `property` cannot import
`resource`. Since Assets and Archives are both types of properties, they must be moved out
of `resource`.

---

Open Question:

This PR moves them to their own sub-folders in `resource`. Should `asset` and `archive`
live somewhere more high level, like `sdk/go/property/{asset,archive}`?
